Refactoring : change folder
