--- src/img/StudyForestLogo.js
@@ -1,8 +1,0 @@
-import React from "react";
-import studyforestlogo from "../img/assets/studyforestlogo.png";
-
-function StudyForestLogo() {
-  return <img src={studyforestlogo} alt="공부의 숲 로고" />;
-}
-
-export default StudyForestLogo;
