--- src/page/TodayHabitPage.js
@@ -1,1 +1,0 @@
-export const TodayHabitPage = () => {};
